move RouteResultTest to context package
